% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsExtract.R
\name{blsExtract}
\alias{blsExtract}
\title{Returns and saves file for BLS data of current month for inds}
\usage{
blsExtract(year, file_path = NULL, create_csv = TRUE)
}
\arguments{
\item{year}{a string for the year to obtain the value}

\item{file}{the file to return winter_wind from}

\item{var2}{2nd wind variable to use}
}
\description{
Takes year for BLS data and returns the values of the different groups for
the BLS employment series file. Requires blsAPI and rjson packages
}
