% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplSeries.R
\name{emplSeries}
\alias{emplSeries}
\title{Returns and saves file for BLS data of current month for inds}
\usage{
emplSeries(
  year,
  file_path = "INVALID FILE PATH",
  file_name = "new_vals",
  create_csv = TRUE
)
}
\arguments{
\item{year}{the year to obtain most recent monthly employment data from}

\item{file_path}{the file path to save the csv file to}

\item{file_name}{The name of the csv file to save}

\item{create_csv}{TRUE to create a CSV file, FALSE to not create one}
}
\description{
Takes year for BLS data and returns the values of the different groups for
the BLS employment series file. Requires blsAPI and rjson packages
}
